+++ milkman-rest/src/main/java/milkman/ui/plugin/rest/postman/importers/PostmanImporterV10.java
@@ -1,9 +1,12 @@
 package milkman.ui.plugin.rest.postman.importers;
 
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
@@ -13,6 +16,7 @@ import com.milkman.rest.postman.schema.v1.PostmanCollection100;
 import com.milkman.rest.postman.schema.v1.Request;
 
 import milkman.domain.Collection;
+import milkman.domain.Folder;
 import milkman.domain.RequestContainer;
 import milkman.ui.plugin.rest.domain.HeaderEntry;
 import milkman.ui.plugin.rest.domain.RestBodyAspect;
@@ -47,17 +51,35 @@ public class PostmanImporterV10 {
 		for(Request container : pmCollection.getRequests()) {
 			requests.addAll(convertToDomain(container));	
 		}
-
-		return new Collection(UUID.randomUUID().toString(), pmCollection.getName(), false, requests);
+		
+		
+//		postman v1.0 doesnt seem to support subfolders, so we just return a list of all folders
+		List<Folder> folders = new LinkedList<>();
+		if (pmCollection.getFolders() != null) {
+			for (com.milkman.rest.postman.schema.v1.Folder postmanFolder : pmCollection.getFolders()) {
+				var folder = convertToDomain(postmanFolder);
+				folders.add(folder);
+			}
+		}
+		return new Collection(UUID.randomUUID().toString(), pmCollection.getName(), false, requests, folders);
 	}
 	
+	private Folder convertToDomain(com.milkman.rest.postman.schema.v1.Folder postmanFolder) {
+		
+		List<String> requests = new LinkedList<>();
+		if (postmanFolder.getOrder() != null)
+			requests.addAll(postmanFolder.getOrder());
+		
+		return new Folder(postmanFolder.getId(), postmanFolder.getName(), new LinkedList(), requests);
+	}
+
+
 	private List<RequestContainer> convertToDomain(Request pmItem) {
 		List<RequestContainer> result = new LinkedList<RequestContainer>();
 		
 		RestRequestContainer request = new RestRequestContainer(pmItem.getName(), pmItem.getUrl(), pmItem.getMethod().toString());
 		
-		//TODO: this should happen automatically...
-		request.setId(UUID.randomUUID().toString());
+		request.setId(pmItem.getId());
 		request.setInStorage(true);
 		
 		//adding headers